Ensure we only have one tracer module per class
In order to prevent the creation of many tracer modules and polluting
the ancestry of a class that grows its traces organically over time, we
can extend the original module with any new traces when we call the
method. This will bust the method cache for that module each time it's
called, but will ultimately be easier to track when using many traces in
a class.