Add tracing for singleton methods
I went with this approach over an approach like the following

```ruby
class Foo
  class << self
    trace_method :foo

    def foo
    end
  end
end
```

because the tracking of what the patchable class was was getting
confusing and making the implementation difficult to follow. This
approach mirrors the structure of `def_method` and
`def_singleton_method` in a way that felt natural to me so I decided to
lean into it.