Move mutant out of the Gemfile
Mutant and Rubocop rely on the same parser gem for parsing so we're
frozen to this version of Rubocop unless we update Mutant. The author of
Mutant decided to make Mutant a commercial product (that is free for
open source projects but requires a registration that can only be done
via an email right now) so we cannot easily update Mutant.

By moving Mutant out of the Gemfile, we remove it as a blocker for
updating Rubocop but can keep using it. Instead of running the `mutant`
task in Rake, now you can use the `bin/mutant` script.